requirements.txt
absl-py==0.8.1
apache-beam==2.23.0
astunparse==1.6.3
attrs==20.1.0
avro-python3==1.9.2.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
contextlib2==0.6.0.post1
crcmod==1.7
cycler==0.10.0
Cython==0.29.21
dataclasses==0.6
dill==0.3.1.1
dm-tree==0.1.5
docopt==0.6.2
fastavro==0.23.6
future==0.18.2
gast==0.3.3
gin-config==0.3.0
google-api-core==1.22.1
google-api-python-client==1.10.1
google-auth==1.20.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-cloud-bigquery==1.27.2
google-cloud-core==1.4.1
google-crc32c==1.0.0
google-pasta==0.2.0
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
grpcio==1.31.0
h5py==2.10.0
hdfs==2.5.8
httplib2==0.17.4
idna==2.10
kaggle==1.5.6
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lxml==4.5.2
Markdown==3.2.2
matplotlib==3.3.1
mock==2.0.0
numpy==1.18.5
oauth2client==3.0.0
oauthlib==3.1.0
object-detection @ file:///C:/Users/Steph/Tensorflow/models-master/research
opencv-python-headless==4.4.0.42
opt-einsum==3.3.0
pandas==1.1.1
pbr==5.4.5
Pillow==7.2.0
promise==2.3
protobuf==3.13.0
psutil==5.7.2
py-cpuinfo==7.0.0
pyarrow==0.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools @ git+https://github.com/philferriere/cocoapi.git@2929bd2ef6b451054755dfd7ceb09278f935f7ad#subdirectory=PythonAPI
pycparser==2.20
pydot==1.4.1
pymongo==3.11.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.14.0
python-slugify==4.0.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
sentencepiece==0.1.91
six==1.15.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-addons==0.11.2
tensorflow-datasets==3.2.1
tensorflow-estimator==2.3.0
tensorflow-hub==0.9.0
tensorflow-metadata==0.23.0
tensorflow-model-optimization==0.4.1
termcolor==1.1.0
text-unidecode==1.3
tf-models-official==2.3.0
tf-slim==1.1.0
tqdm==4.48.2
typeguard==2.9.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.24.3
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
